app/src/test/java/com/maxtyler/adventofcode/puzzles/year2021/day10Test.kt
package com.maxtyler.adventofcode.puzzles.year2021

import org.junit.Assert.assertEquals
import org.junit.Test

class day10Test {
    val input = """
        [({(<(())[]>[[{[]{<()<>>
        [(()[<>])]({[<{<<[]>>(
        {([(<{}[<>[]}>{[]{[(<()>
        (((({<>}<{<{<>}{[]{[]{}
        [[<[([]))<([[{}[[()]]]
        [{[{({}]{}}([{[{{{}}([]
        {<[[]]>}<{[{[{[]{()[[[]
        [<(<(<(<{}))><([]([]()
        <{([([[(<>()){}]>(<<{{
        <{([{{}}[<[[[<>{}]]]>[]]
    """.trimIndent()

    @Test
    fun part1Test() {
        assertEquals("26397", day10.part1(day10.parseInput(input)))
    }

    @Test
    fun part2Test() {
        assertEquals("288957", day10.part2(day10.parseInput(input)))
    }

}